Add files via upload
This is a mini desktop-app  which provides a solution to the problem -lack of availability of proper source of teaching as online dosent provide all the course upto the range of expectation.
situation-I am interested in learning python ,and I have searched all online course which provided me with a certificate but nt a expected knowledge. Through this app we can check for teachers in our locality with affordable price and learn the course with actual guidance
.this will also be a source of pocket money for teachers and development of India.

WORKING APP: first frame-asks for registration of student or teacher with the continuation for teacher registration takes details of teacher and upload it to database .the CHECK OUT option below displays the user details
then we have a basic reference of list of teachers along with their experience price and contact details. we can choose appropriate teacher